Make Ransac use double and not floats

+++ src/Algorithms/Algos/affinemodel.cpp
@@ -7,19 +7,19 @@ AffineModel::AffineModel(const AffineModel&) = default;
 AffineModel::~AffineModel() = default;
 AffineModel& AffineModel::operator=(const AffineModel&) = default;
 
-Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic> AffineModel::getData() const
+Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> AffineModel::getData() const
 {
     return m_A;
 }
 
-void AffineModel::fit(const Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic>& X,
-                      const Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic>& Y)
+void AffineModel::fit(const Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>& X,
+                      const Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>& Y)
 {
     m_A = Y * X.transpose() * (X * X.transpose()).inverse();
 }
 
-Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic>
-AffineModel::predict(const Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic>& X) const
+Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>
+AffineModel::predict(const Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>& X) const
 {
     return m_A * X;
 }
